buttons login and sign up updated but not completely styled yet
