+++ src/App.js
@@ -4,7 +4,7 @@ import './App.css';
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';  
 import LandingPage from './components/Landing_Page/LandingPage.jsx';
 import SignUpPage from './components/Sign_Up_Page/SignUpPage.jsx'
-import SignInPage from './components/Sign_In_Page/SignInPage.jsx'
+import SignInPage from './components/Login_Page/LoginPage.jsx'
 import CardsPage from './components/Cards_Page/CardsPage.jsx'
 
 
@@ -17,7 +17,7 @@ function App() {
         <Routes>
           <Route path="/" element={<LandingPage /> } />
           <Route path="/sign-up" element={<SignUpPage /> } />
-          <Route path="/sign-in" element={<SignInPage />} />
+          <Route path="/log-in" element={<SignInPage />} />
           <Route path="/cards" element={< CardsPage />} />
         </Routes>
       </Router>
