+++ src/components/Login_Page/LoginPage.jsx
@@ -1,0 +1,10 @@
+const LoginPage = () => {
+
+    return (
+        <div>
+            <p>This is the login page</p>
+        </div>
+    );
+};
+
+export default LoginPage;
